+++ aws_account/02_core/variables.tf
@@ -26,9 +26,10 @@ variable "root_domain" {
   default = "example.com"
 }
 
+# https://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_PutRetentionPolicy.html#API_PutRetentionPolicy_RequestSyntax
 variable "log_retention_days" {
   type    = number
-  default = 365
+  default = 400
 }
 
 variable "email" {